Port `struct` snippet
